+++ tests/test_lint.py
@@ -2,6 +2,7 @@ from re import L
 from stac_check.lint import Linter
 import pytest
 
+@pytest.mark.skip(reason="test is ineffective - bad links are redirecting to a third party site")
 def test_linter_bad_asset_requests():
     file = "sample_files/1.0.0/core-item.json"
     linter = Linter(file, assets=True)
@@ -21,8 +22,7 @@ def test_linter_bad_assets():
         "https:/storage.googleapis.com/open-cogs/stac-examples/20201211_223832_CS2.jpg"
     ]
     asset_request_errors = [
-        "https:/storage.googleapis.com/open-cogs/stac-examples/20201211_223832_CS2.jpg",
-        "http://cool-sat.com/catalog/20201211_223832_CS2/20201211_223832_CS2.EPH"
+        "https:/storage.googleapis.com/open-cogs/stac-examples/20201211_223832_CS2.jpg"
     ]
     assert linter.version == "1.0.0"
     assert linter.valid_stac == True
@@ -54,8 +54,7 @@ def test_linter_bad_links_assets():
         "https:/storage.googleapis.com/open-cogs/stac-examples/20201211_223832_CS2.jpg"
     ]
     asset_request_errors = [
-        "https:/storage.googleapis.com/open-cogs/stac-examples/20201211_223832_CS2.jpg",
-        "http://cool-sat.com/catalog/20201211_223832_CS2/20201211_223832_CS2.EPH"
+        "https:/storage.googleapis.com/open-cogs/stac-examples/20201211_223832_CS2.jpg"
     ]
     link_format_errors = ["http:/remotdata.io/catalog/20201211_223832_CS2/index.html"]
     link_request_errors = [
@@ -103,12 +102,121 @@ def test_linter_catalog():
     assert linter.asset_type == "CATALOG"
     assert linter.check_bloated_links() == False
 
-def test_linter_collection_recursive_remote():
-    file = "https://raw.githubusercontent.com/stac-utils/pystac/main/tests/data-files/examples/0.9.0/collection-spec/examples/landsat-collection.json"
+def test_linter_collection_recursive():
+    file = "sample_files/1.0.0/catalog-with-bad-item.json"
     linter = Linter(file, assets=False, links=False, recursive=True)
-    assert linter.version == "0.9.0"
+    assert linter.version == "1.0.0"
     assert linter.recursive == True
-    assert linter.recursive_error_msg == "Exception Could not read uri https://landsat-stac.s3.amazonaws.com/landsat-8-l1/paths/catalog.json"
+    assert linter.validate_all[0] == {
+        "version": "1.0.0",
+        "path": "sample_files/1.0.0/catalog-with-bad-item.json",
+        "schema": [
+            "https://schemas.stacspec.org/v1.0.0/catalog-spec/json-schema/catalog.json"
+        ],
+        "valid_stac": True,
+        "asset_type": "CATALOG",
+        "validation_method": "recursive"
+    }
+
+def test_linter_recursive_max_depth_1():
+    file = "https://radarstac.s3.amazonaws.com/stac/catalog.json"
+    stac = Linter(file, assets=False, links=False, recursive=True, max_depth=1)
+    assert stac.validate_all == [
+        {
+            "version": "0.7.0",
+            "path": "https://radarstac.s3.amazonaws.com/stac/catalog.json",
+            "schema": ["https://cdn.staclint.com/v0.7.0/catalog.json"],
+            "asset_type": "CATALOG",
+            "validation_method": "recursive",
+            "valid_stac": True,
+        }
+    ]
+
+def test_linter_recursive_max_depth_4():
+    file = "https://radarstac.s3.amazonaws.com/stac/catalog.json"
+    stac = Linter(file, assets=False, links=False, recursive=True, max_depth=4)
+    assert stac.validate_all == [
+        {
+            "version": "0.7.0",
+            "path": "https://radarstac.s3.amazonaws.com/stac/catalog.json",
+            "schema": ["https://cdn.staclint.com/v0.7.0/catalog.json"],
+            "asset_type": "CATALOG",
+            "validation_method": "recursive",
+            "valid_stac": True,
+        },
+        {
+            "version": "0.7.0",
+            "path": "https://radarstac.s3.amazonaws.com/stac/radarsat-1/collection.json",
+            "schema": ["https://cdn.staclint.com/v0.7.0/collection.json"],
+            "asset_type": "COLLECTION",
+            "validation_method": "recursive",
+            "valid_stac": True,
+        },
+        {
+            "version": "0.7.0",
+            "path": "https://radarstac.s3.amazonaws.com/stac/radarsat-1/slc/catalog.json",
+            "schema": ["https://cdn.staclint.com/v0.7.0/catalog.json"],
+            "asset_type": "CATALOG",
+            "validation_method": "recursive",
+            "valid_stac": True,
+        },
+        {
+            "version": "0.7.0",
+            "path": "https://radarstac.s3.amazonaws.com/stac/radarsat-1/slc/2012-05-13/RS1_M0630938_F2N_20120513_225708_HH_SLC.json",
+            "schema": ["https://cdn.staclint.com/v0.7.0/item.json"],
+            "asset_type": "ITEM",
+            "validation_method": "recursive",
+            "valid_stac": True,
+        },
+        {
+            "version": "0.7.0",
+            "path": "https://radarstac.s3.amazonaws.com/stac/radarsat-1/slc/2012-06-14/RS1_M0634796_F3F_20120614_110317_HH_SLC.json",
+            "schema": ["https://cdn.staclint.com/v0.7.0/item.json"],
+            "asset_type": "ITEM",
+            "validation_method": "recursive",
+            "valid_stac": True,
+        },
+        {
+            "version": "0.7.0",
+            "path": "https://radarstac.s3.amazonaws.com/stac/radarsat-1/slc/2012-06-14/RS1_M0634795_F3F_20120614_110311_HH_SLC.json",
+            "schema": ["https://cdn.staclint.com/v0.7.0/item.json"],
+            "asset_type": "ITEM",
+            "validation_method": "recursive",
+            "valid_stac": True,
+        },
+        {
+            "version": "0.7.0",
+            "path": "https://radarstac.s3.amazonaws.com/stac/radarsat-1/slc/2012-10-12/RS1_M0634798_F3F_20121012_110325_HH_SLC.json",
+            "schema": ["https://cdn.staclint.com/v0.7.0/item.json"],
+            "asset_type": "ITEM",
+            "validation_method": "recursive",
+            "valid_stac": True,
+        },
+        {
+            "version": "0.7.0",
+            "path": "https://radarstac.s3.amazonaws.com/stac/radarsat-1/slc/2012-10-12/RS1_M0634799_F3F_20121012_110331_HH_SLC.json",
+            "schema": ["https://cdn.staclint.com/v0.7.0/item.json"],
+            "asset_type": "ITEM",
+            "validation_method": "recursive",
+            "valid_stac": True,
+        },
+        {
+            "version": "0.7.0",
+            "path": "https://radarstac.s3.amazonaws.com/stac/radarsat-1/raw/catalog.json",
+            "schema": ["https://cdn.staclint.com/v0.7.0/catalog.json"],
+            "asset_type": "CATALOG",
+            "validation_method": "recursive",
+            "valid_stac": True,
+        },
+        {
+            "version": "0.7.0",
+            "path": "https://radarstac.s3.amazonaws.com/stac/radarsat-1/raw/2012-05-13/RS1_M0000676_F2N_20120513_225701_HH_RAW.json",
+            "schema": ["https://cdn.staclint.com/v0.7.0/item.json"],
+            "asset_type": "ITEM",
+            "validation_method": "recursive",
+            "valid_stac": True,
+        },
+    ]
 
 def test_linter_item_id_not_matching_file_name():
     file = "sample_files/1.0.0/core-item.json"
@@ -170,7 +278,6 @@ def test_title_field():
 def test_self_in_links():
     file = "sample_files/1.0.0/collection-no-title.json"
     linter = Linter(file)
-
     assert linter.check_links_self() == False
     
 def test_catalog_name():